Добавление фичи изменения информации о питомце

+++ PetFamily.Backend/src/PetFamily.Application/Volunteers/Pets/Commands/RemovePet/RemovePetHandler.cs
@@ -1,7 +1,9 @@
 using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.AspNetCore.Server.HttpSys;
 using Microsoft.Extensions.Logging;
 using PetFamily.Application.Abstractions;
+using PetFamily.Application.Extensions;
 using PetFamily.Application.FileProvider;
 using PetFamily.Application.Volunteers.Pets.Commands.AddPet;
 using PetFamily.Domain.Models.VO;
@@ -15,15 +17,24 @@ public class RemovePetHandler : IQueryHandler<string, RemovePetQuery>
 {
     private readonly IFileProvider _fileProvider;
     private readonly ILogger<AddPetHandler> _logger;
+    private readonly IValidator<RemovePetQuery> _validator;
 
-    public RemovePetHandler(IFileProvider fileProvider, ILogger<AddPetHandler> logger)
+    public RemovePetHandler(
+        IFileProvider fileProvider, 
+        ILogger<AddPetHandler> logger, 
+        IValidator<RemovePetQuery> validator)
     {
         _fileProvider = fileProvider;
         _logger = logger;
+        _validator = validator;
     }
 
     public async Task<string> Handle(RemovePetQuery query, CancellationToken cancellationToken)
     {
+        var validationResult = await _validator.ValidateAsync(query, cancellationToken);
+        if (!validationResult.IsValid)
+            return "";
+        
         var filePath = new FilePath(query.FilePath);
         var fileData = new FileInfo(filePath, query.BucketName);
         var deleteFile = await _fileProvider.DeleteFile(fileData, cancellationToken);
